# docs/swagger.yaml
basePath: /
definitions:
...
  auth.OAuthErrorResponse:
    properties:
      error:
        type: string
      error_description:
        type: string
    type: object
  auth.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      id_token: {}
      refresh_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
...
  permissions.CreatePermissionRequest:
    properties:
      description:
        type: string
      key:
        type: string
    required:
    - key
    type: object
  permissions.UpdatePermissionRequest:
    properties:
      description:
        type: string
      key:
        type: string
    type: object
  roles.AssignPermissionsRequest:
    properties:
      permission_ids:
        items:
          type: string
        type: array
    required:
    - permission_ids
    type: object
...
info:
  contact: {}
  description: SSO server API
  title: SSO API
  version: "1.0"
paths:
...
  /auth/authorize:
    get:
      parameters:
      - description: Client ID
        in: query
        name: client_id
        required: true
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        required: true
        type: string
      - description: Scopes (comma-separated)
        in: query
        name: scope
        required: true
        type: string
      - description: State
        in: query
        name: state
        required: true
        type: string
      - description: Provider
        in: query
        name: provider
        type: string
      - description: PKCE code challenge
        in: query
        name: code_challenge
        type: string
      - description: PKCE method (plain or S256)
        in: query
        name: code_challenge_method
        type: string
      - description: OIDC nonce
        in: query
        name: nonce
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
      summary: Authorize
      tags:
      - auth
...
  /auth/login:
    get:
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State
        in: query
        name: state
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
      summary: Login page
      tags:
      - auth
...
  /auth/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: authorization_code or refresh_token
        in: formData
        name: grant_type
        required: true
        type: string
      - description: Source JWT
        in: formData
        name: source_token
        required: true
        type: string
      - description: Authorization code
        in: formData
        name: code
        type: string
      - description: Client ID
        in: formData
        name: client_id
        type: string
      - description: Client secret
        in: formData
        name: client_secret
        type: string
      - description: Redirect URI
        in: formData
        name: redirect_uri
        type: string
      - description: PKCE code verifier
        in: formData
        name: code_verifier
        type: string
      - description: Refresh token
        in: formData
        name: refresh_token
        type: string
      - description: Scopes (comma-separated)
        in: formData
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.OAuthErrorResponse'
      summary: Token
      tags:
      - auth
...
  /permissions:
    get:
      description: List all permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - permissions
    post:
      consumes:
      - application/json
      description: Create a new permission
      parameters:
      - description: Permission creation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/permissions.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create permission
      tags:
      - permissions
  /permissions/{id}:
    delete:
      description: Delete an existing permission
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete permission
      tags:
      - permissions
    get:
      description: Get details of a specific permission
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get permission by ID
      tags:
      - permissions
    put:
      consumes:
      - application/json
      description: Update an existing permission
      parameters:
      - description: Permission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Permission update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/permissions.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update permission
      tags:
      - permissions
...
  /roles/{id}/permissions:
    get:
      description: Get all permissions assigned to a role
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get role permissions
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Assign multiple permissions to a role (replaces existing permissions)
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Permission assignment request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/roles.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign permissions to role
      tags:
      - roles
...
swagger: "2.0"